+++ lib/highs/model.rb
@@ -7,7 +7,7 @@ module Highs
       check_status FFI.Highs_setBoolOptionValue(@ptr, "output_flag", 0)
     end
 
-    def solve(verbose: false)
+    def solve(verbose: false, time_limit: nil)
       num_col = FFI.Highs_getNumCol(@ptr)
       num_row = FFI.Highs_getNumRow(@ptr)
 
@@ -18,7 +18,7 @@ module Highs
       col_basis = IntArray.new(num_col)
       row_basis = IntArray.new(num_row)
 
-      with_options(verbose: verbose) do
+      with_options(verbose: verbose, time_limit: time_limit) do
         check_status FFI.Highs_run(@ptr)
       end
       check_status FFI.Highs_getSolution(@ptr, col_value, col_dual, row_value, row_dual)
@@ -56,11 +56,13 @@ module Highs
       Highs.send(:check_status, status)
     end
 
-    def with_options(verbose:)
+    def with_options(verbose:, time_limit:)
       check_status(FFI.Highs_setBoolOptionValue(@ptr, "output_flag", 1)) if verbose
+      check_status(FFI.Highs_setDoubleOptionValue(@ptr, "time_limit", time_limit)) if time_limit
       yield
     ensure
       check_status(FFI.Highs_setBoolOptionValue(@ptr, "output_flag", 0)) if verbose
+      check_status(FFI.Highs_setDoubleOptionValue(@ptr, "time_limit", Float::INFINITY)) if time_limit
     end
   end
 end